C5 is same to C3, but it uses list instead of intset, and the mutations are performed per host with 1/n prob. A C version probSelect is also added.

+++ c_src/spea2.h
@@ -3,4 +3,4 @@
 
 void spea2_select_2d(double* x_list, double* y_list,
                      size_t num_given,  size_t num_selected,
-                     int* selected);
+                     size_t* selected);
